requirements.txt
absl-py==1.2.0
aiofiles==22.1.0
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.9.2
altgraph==0.17.2
anyio==3.6.1
art==5.7
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.27.0
atlastk==0.13.2
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.11.1
branca==0.5.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cssselect==1.1.0
cssutils==2.6.0
databases==0.6.1
DateTime==4.5
deepface==0.0.75
discord==2.1.0
discord.py==2.1.0
distlib==0.3.6
Django==4.1.1
django-debug-toolbar==3.6.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
dnspython==2.3.0
easyocr==1.5.0
email-validator==1.3.1
emails==0.6
environs==9.5.0
envparse==0.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fake-useragent==0.1.11
fastapi==0.89.1
fastapi-jwt-auth==0.5.0
fibonacci==1.0.0
filelock==3.8.0
fire==0.4.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
flatbuffers==1.12
folium==0.12.1.post1
frozenlist==1.3.0
future==0.18.2
gast==0.4.0
gdown==4.5.1
geojson==2.5.0
gino==1.0.1
gmpy2==2.1.2
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans-py==4.0.0
greenlet==1.1.2
grpcio==1.47.0
gspread==5.4.0
gTTS==2.2.4
h11==0.12.0
h2==4.1.0
h5py==3.7.0
halo==0.0.31
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.3
hyperframe==6.0.1
identify==2.5.15
idna==3.3
image==1.5.33
imageio==2.19.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
log-symbols==0.0.14
lxml==4.9.1
Mako==1.2.3
Markdown==3.4.1
Markups==3.1.3
MarkupSafe==2.1.1
marshmallow==3.17.0
mpmath==1.2.1
mtcnn==0.1.1
multidict==6.0.2
mysqlclient==2.1.1
networkx==2.8.4
nodeenv==1.7.0
numexpr==2.8.1
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
passlib==1.7.4
pdfminer.six==20220524
pdfplumber==0.7.1
pefile==2022.5.30
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.21.0
premailer==3.10.0
protobuf==3.19.4
psycopg==3.1.2
psycopg2==2.9.5
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycoingecko==2.2.0
pycparser==2.21
pydantic==1.10.2
pyfiglet==0.8.post1
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyJWT==1.7.1
pyowm==3.3.0
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.5.1
pytest==7.2.1
pytest-asyncio==0.20.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-markdown-math==0.8
python-multipart==0.0.5
python-weather==0.3.7
pytz==2022.1
PyWavelets==1.3.0
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
qrcode==7.3.1
requests==2.28.0
requests-oauthlib==1.3.1
retina-face==0.0.12
rfc3986==1.5.0
rsa==4.9
scapy==2.4.5
schedule==1.1.0
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
spinners==0.0.24
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.2
starlette==0.22.0
structlog==22.3.0
sympy==1.11.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2022.5.4
tk==0.1.0
tkintertable==1.3.3
tomli==2.0.1
torch==1.12.0
torchvision==0.13.0
tortoise==0.1.1
tqdm==4.64.0
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.9
useragent==0.1.1
uvicorn==0.20.0
virtualenv==20.16.5
vk-api==11.9.9
Wand==0.6.7
watchdog==2.1.9
Werkzeug==2.2.2
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.7.2
zope.interface==5.4.0
